* Cleanup, comments, removed unused stuff * HelpFragment not longer from Markdown * Overlap play now pushes current AudioFile to temp players and replaces it with new one * SoundPlayer now works correctly on config changes mid-play (and mid-pause)

+++ app/src/main/java/us/huseli/soundboard/data/PlayerRepository.kt
@@ -106,7 +106,7 @@ class PlayerRepository @Inject constructor(
         }
     }
 
-    override fun onSoundPlayerDurationChange(sound: Sound, duration: Int) {
+    override fun onSoundPlayerDurationChange(sound: Sound, duration: Long) {
         sound.id?.let { soundId -> soundDao.updateDuration(soundId, duration) }
     }
